use u32 instead of usize to get rid of 'as' conversions

+++ kernel/src/mailbox.rs
@@ -16,16 +16,16 @@ const LFB_MESSAGE_SIZE: usize = 35;
 /// Set physical (display) width/height
 const FB_PHYSICAL_WH_TAG: u32 = 0x00048003;
 /// Width of the requested frame buffer
-const FB_PHYSICAL_WIDTH: u32 = SCREEN_WIDTH as u32;
+const FB_PHYSICAL_WIDTH: u32 = SCREEN_WIDTH;
 /// Height of the requested frame buffer
-const FB_PHYSICAL_HEIGHT: u32 = SCREEN_HEIGHT as u32;
+const FB_PHYSICAL_HEIGHT: u32 = SCREEN_HEIGHT;
 
 pub const FB_BUFFER_LEN: usize = FB_PHYSICAL_HEIGHT as usize * FB_PHYSICAL_WIDTH as usize;
 
 /// Set virtual (buffer) width/height
 const FB_VIRTUAL_WH_TAG: u32 = 0x00048004;
-const FB_VIRTUAL_WIDTH: u32 = SCREEN_WIDTH as u32;
-const FB_VIRTUAL_HEIGHT: u32 = SCREEN_HEIGHT as u32 * 2;
+const FB_VIRTUAL_WIDTH: u32 = SCREEN_WIDTH;
+const FB_VIRTUAL_HEIGHT: u32 = SCREEN_HEIGHT * 2;
 
 pub const TOTAL_FB_BUFFER_LEN: usize = FB_VIRTUAL_HEIGHT as usize * FB_VIRTUAL_WIDTH as usize;
