+++ space_invaders/src/lib.rs
@@ -30,11 +30,11 @@ use crate::framebuffer::Coordinates;
 #[cfg(feature = "std")]
 pub use framebuffer::StdFrameBuffer;
 
-pub const SCREEN_WIDTH: usize = 1280;
-pub const SCREEN_WIDTH_NO_MARGIN: usize = SCREEN_WIDTH - SCREEN_MARGIN;
-pub const SCREEN_HEIGHT: usize = 720;
-pub const SCREEN_HEIGHT_NO_MARGIN: usize = SCREEN_HEIGHT - SCREEN_MARGIN;
-pub const SCREEN_MARGIN: usize = 20;
+pub const SCREEN_WIDTH: u32 = 1280;
+pub const SCREEN_WIDTH_NO_MARGIN: u32 = SCREEN_WIDTH - SCREEN_MARGIN;
+pub const SCREEN_HEIGHT: u32 = 720;
+pub const SCREEN_HEIGHT_NO_MARGIN: u32 = SCREEN_HEIGHT - SCREEN_MARGIN;
+pub const SCREEN_MARGIN: u32 = 20;
 
 pub(crate) const MAX_LIVES: u8 = 3;
